feat: seperating task's domain and data shit to it's own folder
