+++ lib/features/activity/data/activity_repository_impl.dart
@@ -1,9 +1,7 @@
 import 'package:gaia/features/activity/data/datasources/activity_remote_data_source.dart';
 import 'package:gaia/features/activity/data/mappers/exam_mapper.dart';
-import 'package:gaia/features/activity/data/mappers/task_mapper.dart';
 import 'package:gaia/features/activity/domain/activity_repository.dart';
 import 'package:gaia/features/activity/domain/entities/exam_entity.dart';
-import 'package:gaia/features/activity/domain/entities/task_entity.dart';
 import 'package:gaia/features/activity/domain/type/exam_type.dart';
 import 'package:gaia/shared/core/types/result.dart';
 
@@ -23,15 +21,15 @@ class ActivityRepositoryImpl implements ActivityRepository {
         },
       );
 
-  @override
-  Future<Result<List<TaskEntity>>> getTasks({int? page}) => guard(
-        () async {
-          final models = await _dataSource.getTasks(page ?? 1);
-          return models
-              .map(
-                (model) => model.toEntity(),
-              )
-              .toList();
-        },
-      );
+  // @override
+  // Future<Result<List<TaskEntity>>> getTasks({int? page}) => guard(
+  //       () async {
+  //         final models = await _dataSource.getTasks(page ?? 1);
+  //         return models
+  //             .map(
+  //               (model) => model.toEntity(),
+  //             )
+  //             .toList();
+  //       },
+  //     );
 }
