+++ lib/features/activity/data/datasources/activity_remote_data_source.dart
@@ -1,6 +1,5 @@
 import 'package:dio/dio.dart';
 import 'package:gaia/features/activity/data/models/exam_model.dart';
-import 'package:gaia/features/activity/data/models/task_model.dart';
 
 class ActivityRemoteDataSource {
   final Dio _dio;
@@ -25,22 +24,4 @@ class ActivityRemoteDataSource {
         .toList(growable: false);
   }
 
-    Future<List<TaskModel>> getTasks(int page) async {
-    final res = await _dio.get(
-      '/task/all',
-      queryParameters: {
-        'paginate': true,
-        'page' : page,
-        'perPage': 10,
-      },
-    );
-    final data = (res.data as Map<String, dynamic>)['data'] as List<dynamic>;
-    return data
-        .map(
-          (e) => TaskModel.fromJson(
-            Map<String, dynamic>.from(e as Map),
-          ),
-        )
-        .toList(growable: false);
-  }
 }
