+++ lib/features/activity/domain/activity_repository.dart
@@ -1,9 +1,8 @@
 import 'package:gaia/features/activity/domain/entities/exam_entity.dart';
-import 'package:gaia/features/activity/domain/entities/task_entity.dart';
 import 'package:gaia/features/activity/domain/type/exam_type.dart';
 import 'package:gaia/shared/core/types/result.dart';
 
 abstract class ActivityRepository {
   Future<Result<List<ExamEntity>>> getExam(ExamType type, {int? page});
-  Future<Result<List<TaskEntity>>> getTasks({int? page});
+  // Future<Result<List<TaskEntity>>> getTasks({int? page});
 }
