+++ lib/features/activity/presentation/providers/task_controller.dart
@@ -1,7 +1,7 @@
 import 'dart:async';
 import 'package:flutter_riverpod/flutter_riverpod.dart';
-import 'package:gaia/features/activity/domain/entities/task_entity.dart';
-import 'package:gaia/features/activity/presentation/providers/activity_providers.dart';
+import 'package:gaia/features/task/domain/entities/task_entity.dart';
+import 'package:gaia/features/task/presentation/providers/task_providers.dart';
 import 'package:gaia/shared/presentation/paged.dart';
 import 'package:riverpod_annotation/riverpod_annotation.dart';
