+++ lib/features/task/domain/usecase/get_list_task_usecase.dart
@@ -1,9 +1,9 @@
-import 'package:gaia/features/activity/domain/activity_repository.dart';
-import 'package:gaia/features/activity/domain/entities/task_entity.dart';
+import 'package:gaia/features/task/domain/entities/task_entity.dart';
+import 'package:gaia/features/task/domain/task_repository.dart';
 import 'package:gaia/shared/core/types/result.dart';
 
 class GetListTaskUsecase {
-  final ActivityRepository _repository;
+  final TaskRepository _repository;
   GetListTaskUsecase(this._repository);
 
   Future<Result<List<TaskEntity>>> getTasks({int? page}) async {
